+++ src/internal/NeoColors.h
@@ -1,0 +1,44 @@
+/*-------------------------------------------------------------------------
+NeoColors includes all the color classes that describe color and
+modify colors for NeoPixelBus
+
+Written by Michael C. Miller.
+
+I invest time and resources providing this open source code,
+please support me by dontating (see https://github.com/Makuna/NeoPixelBus)
+
+-------------------------------------------------------------------------
+This file is part of the Makuna/NeoPixelBus library.
+
+NeoPixelBus is free software: you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as
+published by the Free Software Foundation, either version 3 of
+the License, or (at your option) any later version.
+
+NeoPixelBus is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public
+License along with NeoPixel.  If not, see
+<http://www.gnu.org/licenses/>.
+-------------------------------------------------------------------------*/
+#pragma once
+
+#include "internal/colors/NeoHueBlend.h"
+
+#include "internal/colors/RgbColorBase.h"
+
+#include "internal/colors/RgbColor.h"
+#include "internal/colors/Rgb16Color.h"
+#include "internal/colors/Rgb48Color.h"
+
+#include "internal/colors/HslColor.h"
+#include "internal/colors/HsbColor.h"
+#include "internal/colors/HtmlColor.h"
+
+#include "internal/colors/RgbwColor.h"
+#include "internal/colors/Rgbw64Color.h"
+
+#include "internal/colors/SegmentDigit.h"
