+++ src/internal/colors/HtmlColor.cpp
@@ -25,6 +25,9 @@ License along with NeoPixel.  If not, see
 -------------------------------------------------------------------------*/
 
 #include <Arduino.h>
+#include "../NeoSettings.h"
+#include "RgbColorBase.h"
+#include "RgbColor.h"
 #include "HtmlColor.h"
 
 static inline char hexdigit(uint8_t v)
