+++ src/internal/colors/HtmlColorNameStrings.cpp
