+++ src/internal/colors/HtmlColorNameStrings.h
