+++ src/internal/colors/HtmlColorNames.cpp
@@ -26,7 +26,10 @@ License along with NeoPixel.  If not, see
 
 
 #include <Arduino.h>
-#include "NeoUtil.h"
+#include "../NeoUtil.h"
+#include "../NeoSettings.h"
+#include "RgbColorBase.h"
+#include "RgbColor.h"
 #include "HtmlColor.h"
 #include "HtmlColorNameStrings.h"
