+++ src/internal/colors/Rgb16Color.h
@@ -1,5 +1,5 @@
 /*-------------------------------------------------------------------------
-RgbColor provides a color object that can be directly consumed by NeoPixelBus
+Rgb16Color provides a color object that stores in only 16 bits, aka 565 format
 
 Written by Michael C. Miller.
 
@@ -25,152 +25,147 @@ License along with NeoPixel.  If not, see
 -------------------------------------------------------------------------*/
 #pragma once
 
-#include "NeoSettings.h"
-#include "RgbColorBase.h"
-
-struct RgbwColor;
-
 // ------------------------------------------------------------------------
-// RgbColor represents a color object that is represented by Red, Green, Blue
-// component values.  It contains helpful color routines to manipulate the 
-// color.
+// Rgb16Color represents a color object that is represented by Red, Green, Blue
+// component values stored in a single 16 bit value using 565 model.  
+// It contains helpful color routines to manipulate the color.
 // ------------------------------------------------------------------------
-struct RgbColor : RgbColorBase
+struct Rgb16Color : RgbColorBase
 {
     typedef NeoRgbCurrentSettings SettingsObject;
 
     // ------------------------------------------------------------------------
-    // Construct a RgbColor using R, G, B values (0-255)
+    // Construct a Rgb16Color using R, G, B values (0-255)
     // ------------------------------------------------------------------------
-    RgbColor(uint8_t r, uint8_t g, uint8_t b) :
-        R(r), G(g), B(b)
+    Rgb16Color(uint8_t r, uint8_t g, uint8_t b) 
     {
+        setR(r);
+        setG(g);
+        setB(b);
     };
 
     // ------------------------------------------------------------------------
-    // Construct a RgbColor using a single brightness value (0-255)
+    // Construct a Rgb16Color using a single brightness value (0-255)
     // This works well for creating gray tone colors
     // (0) = black, (255) = white, (128) = gray
     // ------------------------------------------------------------------------
-    RgbColor(uint8_t brightness) :
-        R(brightness), G(brightness), B(brightness)
+    Rgb16Color(uint8_t brightness) 
+    {
+        setR(brightness);
+        setG(brightness);
+        setB(brightness);
+    };
+
+    // ------------------------------------------------------------------------
+    // Construct a Rgb16Color using a single 16bit encoded value 
+    // CAUTION: the color value must follow a 5/6/5 encoding model otherwise an
+    // incorrect color will be generated
+    // ------------------------------------------------------------------------
+    Rgb16Color(uint16_t color) :
+        Color565(color)
     {
     };
 
     // ------------------------------------------------------------------------
-    // explicitly Construct a RgbColor using RgbwColor
+    // Construct a Rgb16Color using HtmlColor
     // ------------------------------------------------------------------------
-    explicit RgbColor(const RgbwColor& color);
+    Rgb16Color(const HtmlColor& color)
+    {
+        RgbColor converted = color;
+
+        setR(converted.R);
+        setG(converted.G);
+        setB(converted.B);
+    };
 
     // ------------------------------------------------------------------------
-    // Construct a RgbColor using Rgb16Color
+    // Construct a Rgb16Color using HslColor
     // ------------------------------------------------------------------------
-    RgbColor(const Rgb16Color& color);
+    Rgb16Color(const HslColor& color)
+    {
+        RgbColor converted = color;
+
+        setR(converted.R);
+        setG(converted.G);
+        setB(converted.B);
+    };
 
     // ------------------------------------------------------------------------
-    // Construct a RgbColor using HtmlColor
+    // Construct a Rgb16Color using HsbColor
     // ------------------------------------------------------------------------
-    RgbColor(const HtmlColor& color);
+    Rgb16Color(const HsbColor& color)
+    {
+        RgbColor converted = color;
+
+        setR(converted.R);
+        setG(converted.G);
+        setB(converted.B);
+    };
 
     // ------------------------------------------------------------------------
-    // Construct a RgbColor using HslColor
+    // Construct a Rgb16Color that will have its values set in latter operations
+    // CAUTION:  The Color565 member is not initialized and may not be consistent
     // ------------------------------------------------------------------------
-    RgbColor(const HslColor& color);
-
-    // ------------------------------------------------------------------------
-    // Construct a RgbColor using HsbColor
-    // ------------------------------------------------------------------------
-    RgbColor(const HsbColor& color);
-
-
-    // ------------------------------------------------------------------------
-    // Construct a RgbColor that will have its values set in latter operations
-    // CAUTION:  The R,G,B members are not initialized and may not be consistent
-    // ------------------------------------------------------------------------
-    RgbColor()
+    Rgb16Color()
     {
     };
+
+    // ------------------------------------------------------------------------
+    // properties
+    // ------------------------------------------------------------------------
+    void setR(uint8_t r)
+    {
+        Color565 &= 0x07ff;
+        Color565 |= ((r & 0xf8) << 8);
+    };
+    uint8_t getR() const
+    {
+        return (Color565 & 0xf800) >> 8;
+    };
+
+    void setG(uint8_t g)
+    {
+        Color565 &= 0xf81f;
+        Color565 |= ((g & 0xfe) << 3);
+    };
+    uint8_t getG() const
+    {
+        return (Color565 & 0x07e0) >> 3;
+    };
+
+    void setB(uint8_t b)
+    {
+        Color565 &= 0xffe0;
+        Color565 |= ((b & 0xf8) >> 3);
+    };
+    uint8_t getB() const
+    {
+        return (Color565 & 0x001f) << 3;
+    };
+
 
     // ------------------------------------------------------------------------
     // Comparison operators
     // ------------------------------------------------------------------------
-    bool operator==(const RgbColor& other) const
+    bool operator==(const Rgb16Color& other) const
     {
-        return (R == other.R && G == other.G && B == other.B);
+        return (Color565 == other.Color565);
     };
 
-    bool operator!=(const RgbColor& other) const
+    bool operator!=(const Rgb16Color& other) const
     {
         return !(*this == other);
     };
 
     // ------------------------------------------------------------------------
-    // CompareTo method
-    // compares against another color with the given epsilon (delta allowed)
-    // returns the greatest difference of a set of elements, 
-    //   0 = equal within epsilon delta
-    //   negative - this is less than other
-    //   positive - this is greater than other
-    // ------------------------------------------------------------------------
-    int16_t CompareTo(const RgbColor& other, uint8_t epsilon = 1)
-    {
-        return _Compare<RgbColor, int16_t>(*this, other, epsilon);
-    }
-
-    // ------------------------------------------------------------------------
-    // Compare method
-    // compares two colors with the given epsilon (delta allowed)
-    // returns the greatest difference of a set of elements, 
-    //   0 = equal within epsilon delta
-    //   negative - this is less than other
-    //   positive - this is greater than other
-    // ------------------------------------------------------------------------
-    static int16_t Compare(const RgbColor& left, const RgbColor& right, uint8_t epsilon = 1)
-    {
-        return _Compare<RgbColor, int16_t>(left, right, epsilon);
-    }
-
-    // ------------------------------------------------------------------------
-    // operator [] - readonly
-    // access elements in order by index rather than R,G,B
-    // see static Count for the number of elements
-    // ------------------------------------------------------------------------
-    uint8_t operator[](size_t idx) const
-    { 
-        switch (idx)
-        {
-        case 0:
-            return R;
-        case 1:
-            return G;
-        default:
-            return B;
-        }
-    }
-
-    // ------------------------------------------------------------------------
-    // operator [] - read write
-    // access elements in order by index rather than R,G,B
-    // see static Count for the number of elements
-    // ------------------------------------------------------------------------
-    uint8_t& operator[](size_t idx)
-    {
-        switch (idx)
-        {
-        case 0:
-            return R;
-        case 1:
-            return G;
-        default:
-            return B;
-        }
-    }
-
-    // ------------------------------------------------------------------------
     // CalculateBrightness will calculate the overall brightness
     // NOTE: This is a simple linear brightness
     // ------------------------------------------------------------------------
-    uint8_t CalculateBrightness() const;
+    uint8_t CalculateBrightness() const
+    {
+        RgbColor converted = *this;
+        return converted.CalculateBrightness();
+    };
 
     // ------------------------------------------------------------------------
     // Dim will return a new color that is blended to black with the given ratio
@@ -178,7 +173,13 @@ struct RgbColor : RgbColorBase
     // 
     // NOTE: This is a simple linear blend
     // ------------------------------------------------------------------------
-    RgbColor Dim(uint8_t ratio) const;
+    Rgb16Color Dim(uint8_t ratio) const
+    {
+        RgbColor converted = *this;
+        RgbColor result = converted.Dim(ratio);
+
+        return Rgb16Color(result.R, result.G, result.B);
+    };
 
     // ------------------------------------------------------------------------
     // Brighten will return a new color that is blended to white with the given ratio
@@ -186,21 +187,43 @@ struct RgbColor : RgbColorBase
     // 
     // NOTE: This is a simple linear blend
     // ------------------------------------------------------------------------
-    RgbColor Brighten(uint8_t ratio) const;
+    Rgb16Color Brighten(uint8_t ratio) const
+    {
+        RgbColor converted = *this;
+        RgbColor result = converted.Brighten(ratio);
+
+        return Rgb16Color(result.R, result.G, result.B);
+    };
 
     // ------------------------------------------------------------------------
     // Darken will adjust the color by the given delta toward black
     // NOTE: This is a simple linear change
     // delta - (0-255) the amount to dim the color
     // ------------------------------------------------------------------------
-    void Darken(uint8_t delta);
+    void Darken(uint8_t delta)
+    {
+        RgbColor converted = *this;
+
+        converted.Darken(delta);
+        setR(converted.R);
+        setG(converted.G);
+        setB(converted.B);
+    };
 
     // ------------------------------------------------------------------------
     // Lighten will adjust the color by the given delta toward white
     // NOTE: This is a simple linear change
     // delta - (0-255) the amount to lighten the color
     // ------------------------------------------------------------------------
-    void Lighten(uint8_t delta);
+    void Lighten(uint8_t delta)
+    {
+        RgbColor converted = *this;
+
+        converted.Lighten(delta);
+        setR(converted.R);
+        setG(converted.G);
+        setB(converted.B);
+    };
 
     // ------------------------------------------------------------------------
     // LinearBlend between two colors by the amount defined by progress variable
@@ -209,11 +232,21 @@ struct RgbColor : RgbColorBase
     // progress - (0.0 - 1.0) value where 0 will return left and 1.0 will return right
     //     and a value between will blend the color weighted linearly between them
     // ------------------------------------------------------------------------
-    static RgbColor LinearBlend(const RgbColor& left, const RgbColor& right, float progress);
+    static Rgb16Color LinearBlend(const Rgb16Color& left, const Rgb16Color& right, float progress)
+    {
+        RgbColor result = RgbColor::LinearBlend(left, right, progress);
+
+        return Rgb16Color(result.R, result.G, result.B);
+    };
     // progress - (0 - 255) value where 0 will return left and 255 will return right
     //     and a value between will blend the color weighted linearly between them
     // ------------------------------------------------------------------------
-    static RgbColor LinearBlend(const RgbColor& left, const RgbColor& right, uint8_t progress);
+    static Rgb16Color LinearBlend(const Rgb16Color& left, const Rgb16Color& right, uint8_t progress)
+    {
+        RgbColor result = RgbColor::LinearBlend(left, right, progress);
+
+        return Rgb16Color(result.R, result.G, result.B);
+    };
 
     // ------------------------------------------------------------------------
     // BilinearBlend between four colors by the amount defined by 2d variable
@@ -224,54 +257,29 @@ struct RgbColor : RgbColorBase
     // x - unit value (0.0 - 1.0) that defines the blend progress in horizontal space
     // y - unit value (0.0 - 1.0) that defines the blend progress in vertical space
     // ------------------------------------------------------------------------
-    static RgbColor BilinearBlend(const RgbColor& c00, 
-        const RgbColor& c01, 
-        const RgbColor& c10, 
-        const RgbColor& c11, 
-        float x, 
-        float y);
+    static Rgb16Color BilinearBlend(const Rgb16Color& c00,
+        const Rgb16Color& c01,
+        const Rgb16Color& c10,
+        const Rgb16Color& c11,
+        float x,
+        float y)
+    {
+        RgbColor result = RgbColor::BilinearBlend(c00, c01, c10, c11, x, y);
+
+        return Rgb16Color(result.R, result.G, result.B);
+    };
 
     uint32_t CalcTotalTenthMilliAmpere(const SettingsObject& settings)
     {
         auto total = 0;
 
-        total += R * settings.RedTenthMilliAmpere / Max;
-        total += G * settings.GreenTenthMilliAmpere / Max;
-        total += B * settings.BlueTenthMilliAmpere / Max;
+        total += getR() * settings.RedTenthMilliAmpere / 255;
+        total += getG() * settings.GreenTenthMilliAmpere / 255;
+        total += getB() * settings.BlueTenthMilliAmpere / 255;
 
         return total;
-    }
+    };
 
-    // ------------------------------------------------------------------------
-    // Red, Green, Blue color members (0-255) where 
-    // (0,0,0) is black and (255,255,255) is white
-    // ------------------------------------------------------------------------
-    uint8_t R;
-    uint8_t G;
-    uint8_t B;
-
-    const static uint8_t Max = 255;
-    const static size_t Count = 3; // three elements in []
-
-private:
-    inline static uint8_t _elementDim(uint8_t value, uint8_t ratio)
-    {
-        return (static_cast<uint16_t>(value) * (static_cast<uint16_t>(ratio) + 1)) >> 8;
-    }
-
-    inline static uint8_t _elementBrighten(uint8_t value, uint8_t ratio)
-    { 
-        uint16_t element = ((static_cast<uint16_t>(value) + 1) << 8) / (static_cast<uint16_t>(ratio) + 1);
-
-        if (element > Max)
-        {
-            element = Max;
-        }
-        else
-        {
-            element -= 1;
-        }
-        return element;
-    }
+    uint16_t Color565;
 };
