+++ src/internal/colors/Rgb48Color.cpp
@@ -25,8 +25,10 @@ License along with NeoPixel.  If not, see
 -------------------------------------------------------------------------*/
 
 #include <Arduino.h>
-#include "Rgb48Color.h"
+#include "../NeoSettings.h"
+#include "RgbColorBase.h"
 #include "RgbColor.h"
+#include "Rgb48Color.h"
 #include "HslColor.h"
 #include "HsbColor.h"
 #include "HtmlColor.h"
