+++ src/internal/colors/RgbColorBase.cpp
@@ -25,6 +25,7 @@ License along with NeoPixel.  If not, see
 -------------------------------------------------------------------------*/
 
 #include <Arduino.h>
+#include "../NeoSettings.h"
 #include "RgbColorBase.h"
 #include "RgbColor.h"
 #include "Rgb48Color.h"
