+++ src/internal/colors/RgbColorBase.h
