+++ src/internal/colors/Rgbw64Color.cpp
@@ -25,6 +25,8 @@ License along with NeoPixel.  If not, see
 -------------------------------------------------------------------------*/
 
 #include <Arduino.h>
+#include "../NeoSettings.h"
+#include "RgbColorBase.h"
 #include "RgbColor.h"
 #include "RgbwColor.h"
 #include "Rgb48Color.h"
