+++ src/internal/colors/RgbwColor.h
