+++ src/internal/colors/SegmentDigit.h
